fix valid is to slow
Fix the problem that the verification process is too slow

+++ valid.py
@@ -73,7 +73,7 @@ def valid(datacfg, cfgfile, weightfile, outfile):
                 det_conf = box[4]
                 for j in range((len(box)-5)//2):
                     cls_conf = box[5+2*j]
-                    cls_id = box[6+2*j]
+                    cls_id = int(box[6+2*j])
                     prob = det_conf * cls_conf
                     fps[cls_id].write('%s %f %f %f %f %f\n' % (fileId, prob, x1, y1, x2, y2))
